Move some code out of headers
