+++ dataset/include/scipp/dataset/view_decl.h
@@ -106,45 +106,7 @@ private:
 
 public:
   ConstView() = default;
-  ConstView(holder_type &&items, const detail::slice_list &slices = {})
-      : m_items(std::move(items)), m_slices(slices) {
-    // TODO This is very similar to the code in makeViewItems(), provided that
-    // we can give a good definion of the `dims` argument (roughly the space
-    // spanned by all coords, excluding the dimensions that are sliced away).
-    // Remove any items for a non-range sliced dimension. Identified via the
-    // item in case of coords, or via the inner dimension in case of labels and
-    // attributes.
-    for (const auto &s : m_slices) {
-      const auto slice = s.first;
-      if (!slice.isRange()) { // The slice represents a point not a range.
-                              // Dimension removed.
-        for (auto it = m_items.begin(); it != m_items.end();) {
-          auto erase = [slice](const auto &it2) {
-            if constexpr (std::is_same_v<Key, Dim>) {
-              // Remove dimension-coords for given dim, or non-dimension coords
-              // if their inner dim is the given dim.
-              constexpr auto is_dimension_coord = [](const auto &_) {
-                return !is_events(_.second) &&
-                       _.second.dims().contains(_.first);
-              };
-              return is_dimension_coord(it2)
-                         ? it2.first == slice.dim()
-                         : (!it2.second.dims().empty() &&
-                            !is_events(it2.second) &&
-                            (it2.second.dims().inner() == slice.dim()));
-            } else {
-              return !it2.second.dims().empty() && !is_events(it2.second) &&
-                     (it2.second.dims().inner() == slice.dim());
-            }
-          };
-          if (erase(*it))
-            it = m_items.erase(it);
-          else
-            ++it;
-        }
-      }
-    }
-  }
+  ConstView(holder_type &&items, const detail::slice_list &slices = {});
 
   /// Return the number of coordinates in the view.
   index size() const noexcept { return scipp::size(m_items); }
@@ -207,40 +169,10 @@ public:
     return boost::make_transform_iterator(end(), detail::make_value);
   }
 
-  ConstView slice(const Slice slice1) const {
-    auto slices = m_slices;
-    if constexpr (std::is_same_v<Key, Dim>) {
-      const auto &coord = m_items.at(slice1.dim());
-      slices.emplace_back(slice1, coord.dims()[slice1.dim()]);
-    } else {
-      throw std::runtime_error("TODO");
-    }
-    auto items = m_items;
-    return ConstView(std::move(items), slices);
-  }
+  ConstView slice(const Slice slice1) const;
 
-  ConstView slice(const Slice slice1, const Slice slice2) const {
-    return slice(slice1).slice(slice2);
-  }
-  ConstView slice(const Slice slice1, const Slice slice2,
-                  const Slice slice3) const {
-    return slice(slice1, slice2).slice(slice3);
-  }
-
-  bool operator==(const ConstView &other) const {
-    if (size() != other.size())
-      return false;
-    for (const auto &[name, data] : *this) {
-      try {
-        if (data != other[name])
-          return false;
-      } catch (except::NotFoundError &) {
-        return false;
-      }
-    }
-    return true;
-  }
-  bool operator!=(const ConstView &other) const { return !operator==(other); }
+  bool operator==(const ConstView &other) const;
+  bool operator!=(const ConstView &other) const;
 
   const auto &items() const noexcept { return m_items; }
   const auto &slices() const noexcept { return m_slices; }
@@ -273,12 +205,10 @@ private:
 public:
   MutableView() = default;
   MutableView(const Access &access, typename Base::holder_type &&items,
-              const detail::slice_list &slices = {})
-      : Base(std::move(items), slices), m_access(access) {}
+              const detail::slice_list &slices);
   /// Constructor for internal use (slicing and holding const view in mutable
   /// view)
-  MutableView(const Access &access, Base &&base)
-      : Base(std::move(base)), m_access(access) {}
+  MutableView(const Access &access, Base &&base);
 
   /// Return a view to the coordinate for given dimension.
   typename Base::mapped_type::view_type
@@ -325,17 +255,7 @@ public:
 
   MutableView slice(const Slice slice1) const {
     // parent = nullptr since adding coords via slice is not supported.
-    // TODO
     return MutableView(m_access, Base::slice(slice1));
-  }
-
-  MutableView slice(const Slice slice1, const Slice slice2) const {
-    return slice(slice1).slice(slice2);
-  }
-
-  MutableView slice(const Slice slice1, const Slice slice2,
-                    const Slice slice3) const {
-    return slice(slice1, slice2).slice(slice3);
   }
 
   template <class VarOrView>
